# src/main/resources/static/api.yaml
openapi: "3.0.2"
info:
  description: "This is the [Scoold](https://scoold.com) RESTful API. You need to generate an [API key](/admin) from the 'Administration' page to access the API."
  version: "1.0.0"
  title: "Scoold API"
...
paths:
...
  /api/tags:
    post:
      summary: Create a new tag
      tags:
        - tags
      security:
      - scoold_auth: []
      requestBody:
        description: The request entity
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '201':
          description: A new tag object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad request - missing or invalid request entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List tags
      description: Returns a page of tags. Use the `page` parameter to paginate through results.
      tags:
        - tags
      security:
      - scoold_auth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/sortby'
        - $ref: '#/components/parameters/desc'
      responses:
        '200':
          description: A page of tags returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'
  /api/tags/{id}:
    get:
      summary: Get tag
      tags:
        - tags
      security:
      - scoold_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the tag to get
          schema:
            type: string
      responses:
        '200':
          description: Tag object returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: Not found
    patch:
      summary: Rename tag
      tags:
        - tags
      security:
      - scoold_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the tag to rename
          schema:
            type: string
      requestBody:
        description: The request entity
        required: true
        content:
          application/json:
            schema:
              properties:
                tag:
                  description: The new name of the tag
                  type: string
                  example: new-tag-name
      responses:
        '200':
          description: Tag was renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: Not found
    delete:
      summary: Delete tag
      tags:
        - tags
      security:
      - scoold_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the tag to delete
          schema:
            type: string
      responses:
        '200':
          description: Tag was deleted
        '404':
          description: Not found
...
  /api/reports:
    post:
      summary: Create a new report
      description: You can store any additional properties in the report object - simply pass those as JSON in the request body.
      tags:
        - reports
      security:
      - scoold_auth: []
      requestBody:
        description: Set `creatorid` to change the author of the report.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Report'
      responses:
        '201':
          description: A new report object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Bad request - missing or invalid request entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List reports
      tags:
        - reports
      security:
      - scoold_auth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/sortby'
        - $ref: '#/components/parameters/desc'
      responses:
        '200':
          description: a page of report objects returned
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Reports'
...
  /api/reports/{id}/close:
    put:
      summary: Close report
      tags:
        - reports
      security:
      - scoold_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the report to close
          schema:
            type: string
      requestBody:
        description: The request entity
        required: true
        content:
          application/json:
            schema:
              properties:
                solution:
                  description: Actions taken to solve the issue
                  type: string
      responses:
        '200':
          description: Report was closed
...
components:
  schemas:
...
    Tag:
      type: object
      properties:
        tag:
          type: string
        count:
          type: integer
    Tags:
      type: array
      items:
        $ref: '#/components/schemas/Tag'
...
    Report:
      type: object
      properties:
        id:
          type: string
        creatorid:
          type: string
        subType:
          type: string
          enum:
            - SPAM
            - OFFENSIVE
            - DUPLICATE
            - INCORRECT
            - OTHER;
        timestamp:
          type: integer
          format: int64
        solution:
          type: string
        link:
          type: string
        closed:
          type: boolean
        authorName:
          type: string
    Reports:
      type: array
      items:
        $ref: '#/components/schemas/Report'
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  parameters:
    limit:
      name: limit
      in: query
      description: How many items to return at one time (default 30)
      required: false
      schema:
        type: integer
        format: int32
        default: 30
    page:
      name: page
      in: query
      description: Page number (paginate through results)
      required: false
      schema:
        type: integer
        format: int32
        default: 1
    sortby:
      name: sortby
      in: query
      description: Field name on which results will be sorted
      required: false
      schema:
        type: string
        default: "timestamp"
    desc:
      name: desc
      in: query
      description: Descending (true) or ascending (false) sort order
      required: false
      schema:
        type: boolean
        default: true
  securitySchemes:
    scoold_auth:
      description: Don't have an API key? Generate one on the [Administration page](/admin).
      type: http
      scheme: bearer
      bearerFormat: JWT